requirement.txt
Flask-Caching
Flask-MySQLdb
Flask
Flask-Session
Flask-MySQL
pycryptodome